Update class.add-from-server.php
My last pull request was not complete, $time is used elsewhere as timestamp. I've added the translation to YYYY/MM format only where needed.
Furthermore all images are added with GMT as post time, which makes it difficult to find the images in the media library. I changed the post_date to use the current time and time zone.